Move `API::BaseController` to `APIController`
To be consistent with `AdminController` and `SchoolsController`.
