+++ app/controllers/api/v3/transfers_controller.rb
@@ -1,6 +1,6 @@
 module API
   module V3
-    class TransfersController < BaseController
+    class TransfersController < APIController
       def index = head(:method_not_allowed)
       def show = head(:method_not_allowed)
     end
