+++ app/controllers/api/v3/unfunded_mentors_controller.rb
@@ -1,6 +1,6 @@
 module API
   module V3
-    class UnfundedMentorsController < BaseController
+    class UnfundedMentorsController < APIController
       def index = head(:method_not_allowed)
       def show = head(:method_not_allowed)
     end
